Render core reworks
