+++ CorgEng.Core/CorgEng.Core.sln
@@ -34,8 +34,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorgEng.ContentLoading", ".
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorgEng.MapLoading", "..\CorgEng.MapLoading\CorgEng.MapLoading.csproj", "{B324FB97-A466-4E19-BA67-9536023EE5B9}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutoMap", "..\AutoMap\AutoMap.csproj", "{ADA9CC94-5705-499C-9C93-888E96F87DEA}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorgEng.UserInterface", "..\CorgEng.UserInterface\CorgEng.UserInterface.csproj", "{19E048D9-1587-40CC-BEE3-E1883A21FEC8}"
 	ProjectSection(ProjectDependencies) = postProject
 		{BFC5583F-3A05-43E4-9D07-81D80B1EC685} = {BFC5583F-3A05-43E4-9D07-81D80B1EC685}
@@ -331,26 +329,6 @@ Global
 		{B324FB97-A466-4E19-BA67-9536023EE5B9}.Release|x64.Build.0 = Release|x64
 		{B324FB97-A466-4E19-BA67-9536023EE5B9}.Release|x86.ActiveCfg = Release|x86
 		{B324FB97-A466-4E19-BA67-9536023EE5B9}.Release|x86.Build.0 = Release|x86
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|ARM32.ActiveCfg = Debug|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|ARM32.Build.0 = Debug|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|x64.ActiveCfg = Debug|x64
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|x64.Build.0 = Debug|x64
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|x86.ActiveCfg = Debug|x86
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Debug|x86.Build.0 = Debug|x86
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|ARM32.ActiveCfg = Release|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|ARM32.Build.0 = Release|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|ARM64.Build.0 = Release|Any CPU
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|x64.ActiveCfg = Release|x64
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|x64.Build.0 = Release|x64
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|x86.ActiveCfg = Release|x86
-		{ADA9CC94-5705-499C-9C93-888E96F87DEA}.Release|x86.Build.0 = Release|x86
 		{19E048D9-1587-40CC-BEE3-E1883A21FEC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{19E048D9-1587-40CC-BEE3-E1883A21FEC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19E048D9-1587-40CC-BEE3-E1883A21FEC8}.Debug|ARM32.ActiveCfg = Debug|Any CPU
