+++ CorgEng.Core/CorgEngMain.cs
@@ -6,6 +6,8 @@ using CorgEng.Core.Rendering;
 using CorgEng.Core.Rendering.Exceptions;
 using CorgEng.GenericInterfaces.Logging;
 using CorgEng.GenericInterfaces.Rendering;
+using CorgEng.GenericInterfaces.Rendering.Renderers;
+using CorgEng.GenericInterfaces.Rendering.Renderers.SpriteRendering;
 using GLFW;
 using System;
 using System.Collections.Concurrent;
@@ -24,6 +26,12 @@ namespace CorgEng.Core
     {
 
         /// <summary>
+        /// Get a dependency for a default sprite renderer
+        /// </summary>
+        [UsingDependency]
+        private static ISpriteRendererFactory SpriteRendererFactory = null!;
+
+        /// <summary>
         /// The internal render master.
         /// Handles transfering iamges generated by the RenderCore to 
         /// the screen.
@@ -35,18 +43,18 @@ namespace CorgEng.Core
         /// If none is provided then a renderer provided one will be given which just draws things on its plane and performs
         /// no other special effects.
         /// </summary>
-        private static ConcurrentDictionary<int, RenderCore> renderCores { get; } = new ConcurrentDictionary<int, RenderCore>();
+        private static ConcurrentDictionary<int, IRenderer> renderCores { get; } = new ConcurrentDictionary<int, IRenderer>();
 
         /// <summary>
         /// Render planes that are staged to start rendering
         /// </summary>
-        private static SortedList<int, RenderCore> stagedRenderPlanes = new SortedList<int, RenderCore>();
+        private static SortedList<int, IRenderer> stagedRenderPlanes = new SortedList<int, IRenderer>();
 
         /// <summary>
         /// A sorted list containing the planes that are currently drawing directly to the screen.
         /// Planes are drawn in order of layer when they aren't drawn to another surface first.
         /// </summary>
-        private static SortedList<int, RenderCore> renderingPlanes = new SortedList<int, RenderCore>();
+        private static SortedList<int, IRenderer> renderingPlanes = new SortedList<int, IRenderer>();
 
         /// <summary>
         /// The window associated with the CorgEng application
@@ -201,7 +209,7 @@ namespace CorgEng.Core
                 foreach (var renderPlane in renderingPlanes.Values)
                 {
                     // Process the actual render
-                    renderPlane.DoRender();
+                    renderPlane.Render(MainCamera);
 					//Pass the output image from the render core to the internal renderer
 					InternalRenderMaster.RenderImageToScreen(renderPlane);
 				}
@@ -230,7 +238,7 @@ namespace CorgEng.Core
         /// Resize all of the contained render cores when the screen size changes
         /// </summary>
         private static CorgEngWindow.WindowResizeDelegate activeSizeDelegate = (width, height) => {
-            foreach (RenderCore renderCore in renderCores.Values)
+            foreach (IRenderer renderCore in renderCores.Values)
             {
                 renderCore.Resize(width, height);
             }
@@ -239,7 +247,7 @@ namespace CorgEng.Core
         /// <summary>
         /// Sets the CorgEng program's render core.
         /// </summary>
-        public static void SetPlaneRenderCore(int renderCorePlane, RenderCore newRenderCore)
+        public static void SetPlaneRenderCore(int renderCorePlane, IRenderer newRenderCore)
         {
 			newRenderCore.Initialize();
 			renderCores.AddOrUpdate(renderCorePlane, newRenderCore, (plane, oldCore) => {
@@ -255,12 +263,14 @@ namespace CorgEng.Core
         /// <returns></returns>
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static RenderCore GetRenderCore(int renderCorePlane)
+		public static IRenderer GetRendererForPlane(int renderCorePlane)
         {
             return renderCores.GetOrAdd(renderCorePlane, plane => {
                 // Fetch a default render core implementation
-                //return new RenderCore();
-            });
+                var spriteRenderer = SpriteRendererFactory.CreateSpriteRenderer(plane);
+                stagedRenderPlanes.Add(plane, spriteRenderer);
+				return spriteRenderer;
+			});
 		}
 
         /// <summary>
