--- CorgEng.Rendering/RenderCores/DefaultRenderCore.cs
@@ -1,27 +1,0 @@
-﻿using CorgEng.Core.Rendering;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CorgEng.Rendering.RenderCores
-{
-	internal class DefaultRenderCore : RenderCore
-	{
-		public override void Dispose()
-		{
-			throw new NotImplementedException();
-		}
-
-		public override void Initialize()
-		{
-			throw new NotImplementedException();
-		}
-
-		public override void PerformRender()
-		{
-			throw new NotImplementedException();
-		}
-	}
-}
